Apply Safe Autofixes Clippy

+++ src-tauri/src/inpainting.rs
@@ -378,8 +378,8 @@ fn calculate_normal(pixel_states: &[u8], width: u32, height: u32, x: u32, y: u32
         }
     };
 
-    let grad_x = (state_at(x_p1, y) as i32 - state_at(x_m1, y) as i32) as f32;
-    let grad_y = (state_at(x, y_p1) as i32 - state_at(x, y_m1) as i32) as f32;
+    let grad_x = (state_at(x_p1, y) - state_at(x_m1, y)) as f32;
+    let grad_y = (state_at(x, y_p1) - state_at(x, y_m1)) as f32;
     let mag = (grad_x * grad_x + grad_y * grad_y).sqrt();
     if mag > 1e-6 {
         (-grad_y / mag, grad_x / mag)
